Added SSL_CTX_get0_certificate()
