+++ bssl-compat/source/extra/x509_extra.ld
@@ -1,0 +1,2 @@
+PROVIDE(X509_cmp = ossl_X509_cmp);
+PROVIDE(X509_free = ossl_X509_free);
